Add basic smoke test job on CircleCI (#2)
Ideally we'd do the build on Circle, then trigger the smoke test after, but I did the build on GH actions because I was lazy. Maybe I'll fix that someday. For now, this just runs a job on Circle using the image, and does a couple basic checks to make sure things work.