Support Python 3.15.0a2, allow broken pipenv/pipx (#35)
Pipenv is currently broken on Python 3.15.0a2, so this updates the Dockerfile and build process to allow Pipenv to be left out of the image (just like we do for Poetry). This does the same for Pipx in case it has similar issues in the future, even though works fine in this release.